refac(config): fix imports and moved version to pkg/event where it is used.
