+++ pkg/client/fixtures_test.go
@@ -19,8 +19,8 @@ package client
 
 import (
 	"fmt"
-	"github.com/optimizely/go-sdk/pkg/config"
 
+	"github.com/optimizely/go-sdk/pkg/config"
 	"github.com/optimizely/go-sdk/pkg/decision"
 	"github.com/optimizely/go-sdk/pkg/entities"
 	"github.com/optimizely/go-sdk/pkg/event"
